backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hackathon-todo"
version = "0.2.0"
description = "Multi-user todo web application backend API with FastAPI"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Development Team", email = "dev@example.com"}
]

dependencies = [
    "fastapi==0.109.0",
    "sqlmodel==0.0.14",
    "alembic==1.13.1",
    "pydantic-settings==2.1.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-dotenv==1.0.0",
    "uvicorn[standard]==0.27.0",
    "sqlalchemy>=2.0.28",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0",
    "email-validator>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "httpx==0.25.2",
    "black==23.12.0",
    "flake8==6.1.0",
    "mypy==1.7.1",
    "isort==5.13.2",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.isort]
profile = "black"
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-v --strict-markers --tb=short"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
